extract len check function to base
